+++ assistant/src/runtime/auth/__tests__/guard-tests.test.ts
@@ -66,7 +66,6 @@ describe("route policy coverage", () => {
       "browser-extension-pair-routes.ts",
       "guardian-bootstrap-routes.ts",
       "guardian-refresh-routes.ts",
-      "meet-internal.ts",
     ]);
     const allSources = [httpServerSrc];
     try {
